% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{add_error}
\alias{add_error}
\title{Add error to diploid}
\usage{
add_error(diploid, hom.error, het.error)
}
\arguments{
\item{diploid}{a vector of diploid genotype, for example, c(0, 1, 1, 0, 2, 2).}

\item{hom.error}{homozygous error rate.}

\item{het.error}{heterozygous error rate.}
}
\value{
a vector of diploid genotype.
}
\description{
Add error to diploid
}
\examples{
add_error(c(0, 1, 1, 0, 2, 2), 0.02, 0.8)
}

