% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{copy.mom}
\alias{copy.mom}
\title{Copy mom to kids with recombination}
\usage{
copy.mom(mom, co_mean)
}
\arguments{
\item{mom}{a list of mom's haplotype.}

\item{co_mean}{mean number of crossovers per chromosome.}
}
\value{
a vector of diploid genotype.
}
\description{
Copy mom to kids with recombination
}
\examples{
copy.mom(list(c(0, 1, 1, 0, 1, 1)), 0.25)
}

