% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_array.R
\name{create_array}
\alias{create_array}
\alias{est_maf}
\alias{est_missing}
\alias{est_pop_maf}
\alias{get_array_item}
\alias{get_snpinfo}
\alias{ped_focal}
\alias{pedinfo}
\title{\code{Create GBS.array objects}}
\usage{
create_array(geno, ped, pargeno, pinfo = NULL, snpinfo = NULL,
  outdir = "largedata/obs", bychr = FALSE, self_cutoff = 30)

get_snpinfo(geno, ped, self_cutoff)

pedinfo(ped)

est_pop_maf(geno, pinfo, ped, self_cutoff)

est_maf(pgeno)

est_missing(pgeno)

ped_focal(ped, pargeno, focalp)

get_array_item(subgeno, myped, focalp)
}
\arguments{
\item{geno}{a data.frame of gentoype. coded with 0 (major), 1 (het), 2 (minor) and 3 (missing).}

\item{ped}{a data.frame of pedigree information, must include three columns: proid (progeny id),
parent1 (the first parent id), parent2 (the 2nd parent id).}

\item{pargeno}{A data.frame with two columns: "parentid" and "true_p".
The 2nd column indicates whether the true parental genotype is available.}

\item{snpinfo}{Wether to calculate snpinfo (this data.frame can be obtained by using function of get_snpinfo).}

\item{outdir}{path of the output dir.}

\item{bychr}{Wether to split the data by 10 chromsomes.}

\item{size_cutoff}{Minimum family size required for calculate snpinfo, default=30.}
}
\value{
Return updated Geno4imputeR object.
The GBS.array objects will be output to the given directory by family and chr.

Details of the GBS.array objects.
Slot1: true_parents, a list of data.frame(hap1, hap2).
Slot2: gbs_parents, a list of genotypes. For example, c(1, 2, 2, 0, 3).
Slot3: true_kids, a list of data.frame(hap1, hap2).
Slot4: gbs_kids, a list of kid genotypes. For example, c(1, 1, 3, 1, 2).
Slot5: pedigree, a data.frame (kid, p1, p2). Note, p1 is the focal parent.
Slot6: freq, a vector of reference allele freq for all SNPs.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
Create GBS.array objects from Geno4imputeR objects. It will compute population allele frq based on families
of selfed kids. Original Geno4imputeR will be updated and returned. A set of GBS.array objects will be output
to the directory by chromosomes.
}
\examples{
create_array(Geno4imputeR, ped, outdir="largedata/",
maf_cutoff=0.002, lmiss_cutoff=0.8, imiss_cutoff=0.8, size_cutoff=40)
}

