% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_missing_values.R
\name{detect_missing_values}
\alias{detect_missing_values}
\title{Detect Missing Values in a Dataset}
\usage{
detect_missing_values(data)
}
\arguments{
\item{data}{A data frame or matrix.}
}
\value{
A named vector with the count of missing values for each column that has missing data.
}
\description{
This function detects and counts the number of missing values for each column in a dataset.
}
\examples{
sample_data <- data.frame(
  A = c(1, 2, NA, 4, 5),
  B = c(NA, 2, 3, 4, 5),
  C = c(1, 2, 3, 4, 5)
)
detect_missing_values(sample_data)
}
