% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{error_mx}
\alias{error_mx}
\alias{gen_error_mat}
\title{Return error matrices}
\usage{
error_mx(hom.error, het.error)

gen_error_mat(hom.error, het.error)
}
\arguments{
\item{hom.error}{homozygous error rate, default=0.02.}

\item{het.error}{heterozygous error rate, default=0.8.}
}
\value{
\code{error_mx} returns a list of three matrices.
In this matrix,
row1 is true_gen 00, row2 is true_gen 01, row3 is true_gen 11.
cols 1-3 are observed genotype (00,01,11) and last col4 is the missing data.
1. 00x01
2. 01x01
3. 11x01
}
\description{
Return error matrices
}
\examples{
#Genotype error matrix
gen_error_mat(hom.error=0.02, het.error=0.8)

#Genotype error by considering Mendelian segregation rate
error_mx(hom.error=0.02, het.error=0.8)
}

