% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{error_probs}
\alias{error_mx2}
\alias{error_probs}
\alias{gen_error_mat}
\title{Return error matrices}
\usage{
error_probs(mx, merr)

gen_error_mat(major.error, het.error, minor.error)

error_mx2(major.error, het.error, minor.error)
}
\arguments{
\item{mx}{A error matrix.}

\item{merr}{Error value for missing data.}
}
\value{
\code{error_mx} returns a list of three matrices.
In this matrix,
row1 is true_gen 00, row2 is true_gen 01, row3 is true_gen 11.
cols 1-3 are obs. genotype (00,01,11) and last col4 is the missing data.
1. 00x01
2. 01x01
3. 11x01
}
\description{
Return error matrices
}
\examples{
#Genotype error matrix
gen_error_mat(hom.error=0.02, het.error=0.8)

#Genotype error by consideringMendelian segregation rate
error_mx(hom.error=0.02, het.error=0.8)
}

