% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_error.R
\name{estimate_error}
\alias{check_error}
\alias{count_geno}
\alias{estimate_error}
\alias{kid_het_err}
\title{\code{Estimate genotyping error}}
\usage{
estimate_error(geno, ped, self_cutoff, depth_cutoff, est_kids = FALSE)

kid_het_err(geno, depth_cutoff, ped, pinfo, verbose)

count_geno(subgeno)

check_error(subgeno, depth_cutoff)
}
\arguments{
\item{geno}{input object.}

\item{ped}{a data.frame of pedigree information, must include three columns: proid (progeny id),
parent1 (the first parent id), parent2 (the 2nd parent id).}

\item{imiss_cutoff}{cutoff for individual missing rate, default=0.8.}

\item{size_cutoff}{Minimum family size required for imputation, default=40.}
}
\value{
Return error matrix.
  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
Estimate error from data and pedigree.
}
\examples{
library(data.table, lib="~/bin/Rlib/")
ped <- read.table("data/parentage_info.txt", header =TRUE)
ped[, 1:3] <- apply(ped[, 1:3], 2, as.character)
geno <- fread("largedata/lcache/teo_recoded.txt")
estimate_error(geno, ped, self_cutoff=30, depth_cutoff=10)
}

