% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GBS.array-class}
\alias{GBS.array-class}
\title{A S4 class stores genomic data.}
\description{
Stores GBS data of progeny array and parental data, as well as pedigree information.
In the GBS data, 0 ref, 1 het, 2 alt and 3 missing.
}
\section{Slots}{

\describe{
\item{\code{true_parents}}{List of parents' read data. List index must consistent with pedigree.}

\item{\code{gbs_parents}}{List of parents' GBS data.}

\item{\code{true_kids}}{List of progeny's readl data.}

\item{\code{gbs_kids}}{List of progeny's GBS data.}

\item{\code{pedigree}}{Pedigree information.}
}}

