% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_phased}
\alias{get_phased}
\title{\code{Method for GBS.array}}
\usage{
get_phased(GBS.array, maxchunk)
}
\arguments{
\item{GBS.array}{Input a GBS.array object.}

\item{maxchunk}{The max number of chunks.}
}
\description{
Simulate phased parents. All the gbs_parents are perfectly phased.
}

