% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{getsfs}
\alias{getsfs}
\title{setup the neutral SFS}
\usage{
getsfs(x = 1:99/100)
}
\arguments{
\item{x}{A vector of freq bins. Default 0.01-0.99.}
}
\value{
Return a vector of neutral SFS (N=51734)
}
\description{
setup the neutral SFS
}
\examples{
sfs <- getsfs(x=1:99/100)
hist(sfs)
}

