% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{hw_probs}
\alias{hw_probs}
\title{Return HW probs}
\usage{
hw_probs(x)
}
\arguments{
\item{x}{Allele freq p.}
}
\value{
Return a vector of p^2, 2pq and q^2,
}
\description{
Return HW probs
}
\examples{
hw_probs(0.1)
}

