\name{impute-package}
\alias{impute-package}
\docType{package}
\title{
Impute missing values by automated variable selection
}
\description{
The function 'impute' in this package is the main function that uses variable selection to impute missing values. Many different variable selection methods have been incorporated in this package and they are listed in the mannual. 
}
\details{
\tabular{ll}{
Package: \tab impute\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-14\cr
License: GPL (>= 2)
}
The function 'impute' in this package is the main function that uses variable selection to impute missing values. Many different variable selection methods have been incorporated in this package and they are listed in the mannual. 
}
\author{
Lingbing Feng
Maintainer: Lingbing Feng <fenglb88@gmail.com>

}

\keyword{imputation, variable selection}
\examples{
a <- 1
}
