% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImputeKid.R, R/imputeKids.R
\name{impute_kid}
\alias{chr_hap}
\alias{hap_in_chunk}
\alias{impute_kid}
\alias{which_kid_hap}
\title{\code{Phasing and Imputing Kid Genotypes}}
\usage{
impute_kid(GBS.array, winstart = 10, winend = 500, stepsize = 10,
  recomb = 1.5, verbose = TRUE)

chr_hap(GBS.array, k, winsize)

hap_in_chunk(momphase, c, win_length, kid)

which_kid_hap(haplotype, kidwin)

impute_kid(GBS.array, winstart = 10, winend = 500, stepsize = 10,
  recomb = 1.5, verbose = TRUE)
}
\arguments{
\item{GBS.array}{A GBS.array object with phased parents.
It can be generated from \code{phase_parent} or\code{sim.array} functions.}

\item{verbose}{Writing verbose messages. Default=TRUE.}

\item{win_length}{Window length used for halotype phasing. Default=10.
win_length > 20 will dramatically increase computational burden.}

\item{kidwin}{The genotypes of the kid for all heterozygote site of both parents.}

\item{join_length}{The length of each neighboring chunks used to connect them into a longer one.}

\item{p1_haps}{Two haplotypes in a chunk of the first parent.}

\item{p2_haps}{All possible haplotypes in chunks of the 2nd parent.}

\item{GBS.array}{A GBS.array object with phased parents.
It can be generated from \code{phase_parent} or\code{sim.array} functions.}

\item{win_length}{Window length used for halotype phasing. Default=10.
win_length > 20 will dramatically increase computational burden.}

\item{join_length}{The length of each neighboring chunks used to connect them into a longer one.}

\item{verbose}{Writing verbose messages. Default=TRUE.}
}
\value{
return a data.frame with all the log likelihoods. Using function \code{momgeno} to extract mom's genotype.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.

return a data.frame with all the log likelihoods. Using function \code{momgeno} to extract mom's genotype.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
Calculate the maximum likelihood of the inherited haplotypes. And then find the minimum path of recombinations.
At the potential recombination sites, find the maximum likelihood of the breaking points.

Calculate the maximum likelihood of the inherited haplotypes. And then find the minimum path of recombinations.
At the potential recombination sites, find the maximum likelihood of the breaking points.
}
\examples{
GBS.array <- sim.array(size.array=50, numloci=1000, imiss=0.2, selfing=0.5)
GBS.array <- sim.array(size.array=50, numloci=1000, imiss=0.2, selfing=0.5)
}

