% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImputeMom.R
\name{impute_mom}
\alias{impute_mom}
\title{\code{Impute mom's genotype using mom and kids' GBS data. }}
\usage{
impute_mom(obs_mom, obs_kids, locus, p)
}
\arguments{
\item{obs_mom}{A vector of mom's GBS data. Coded with 0, 1 and 2, the copy of reference alleles.
Missing data should be coded with 3.}

\item{obs_kids,}{A logical scalar. Should missing values (including NaN)
be removed?}

\item{p}{p=sample(sfs,numloci) #freqs of all loci sampled from site frequency spectrum (SFS)}
}
\value{
If all inputs are integer and logical, then the output
  will be an integer. If integer overflow
  \url{http://en.wikipedia.org/wiki/Integer_overflow} occurs, the output
  will be NA with a warning. Otherwise it will be a length-one numeric or
  complex vector.

  Zero-length vectors have sum 0 by definition. See
  \url{http://en.wikipedia.org/wiki/Empty_sum} for more details.
}
\description{
We have obs. mom and obs. (selfed) kids. We want to know the most likelihood of the mom's genotype.
This function is to impute mom's genotype from a progeny array of k kids at a single locus.
inferred_mom=1 -> 00, 2->01, 3->11
}
\examples{
sum(1:10)
sum(1:5, 6:10)
sum(F, F, F, T, T)

sum(.Machine$integer.max, 1L)
sum(.Machine$integer.max, 1)

\dontrun{
sum("a")
}
}

