% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imputeKids.R
\name{impute_onekid}
\alias{impute_onekid}
\title{\code{Impute kid genotypes}}
\usage{
impute_onekid(momphase, kid, win_length, verbose)
}
\arguments{
\item{win_length}{Window length used for halotype phasing. Default=10.
win_length > 20 will dramatically increase computational burden.}

\item{verbose}{Writing verbose messages. Default=TRUE.}

\item{GBS.array}{A GBS.array object with phased parents.
It can be generated from \code{phase_parent} or\code{sim.array} functions.}

\item{join_length}{The length of each neighboring chunks used to connect them into a longer one.}
}
\value{
return a data.frame with all the log likelihoods. Using function \code{momgeno} to extract mom's genotype.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
Calculate the maximum likelihood of the inherited haplotypes.
}
\examples{
GBS.array <- sim.array(size.array=50, numloci=1000, imiss=0.2, selfing=0.5)
}

