\name{imputer-package}
\alias{imputer-package}
\alias{imputer}
\docType{package}
\title{
\packageTitle{imputer}
}
\description{
\packageDescription{imputer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{imputer}
\packageIndices{imputer}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{imputer}

Maintainer: \packageMaintainer{imputer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
