% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PhaseParent.R
\name{join_chunks}
\alias{join_chunks}
\alias{link_dad_haps}
\title{\code{Joining phased haplotypes chunks. }}
\usage{
join_chunks(GBS.array, chunklist, verbose, join_len)

link_dad_haps(GBS.array, dad_haps_lofl, hapidx, join_len)
}
\arguments{
\item{GBS.array}{A GBS.array object generated from create_GBS.array() or sim.array() functions.}

\item{chunklist}{A list of haplotype chunks.}

\item{verbose}{For function 'genomom', returnall=TRUE will return with all the information;
returnall=FALSE will return
mom's genotype only.}

\item{join_length}{A list of vectors of Kid's GBS data.
Coded with 0, 1 and 2, which is the copy of alternative alleles.
Missing data should be coded with 3.}
}
\value{
return a data.frame with all the log likelihoods.
Using function \code{momgeno} to extract mom's genotype.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
Extending/joining the phased chunks of each chromosome generated by \code{phase_chunk}.
}
\examples{
GBS.array <- sim.array(size.array=50, numloci=1000)
haps <- setup_haps(win_length)
chunklist <- phase_chuck(GBS.array, haps, verbose, win_length)
haplist <- join_chunks(GBS.array, chunklist, verbose, join_len)
}

