% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{kid}
\alias{kid}
\title{Returns a list of true [[1]] and observed [[2]] kid}
\usage{
kid(p1, p2, het.error, hom.error, rec = 1.5, imiss = 0, misscode = 3)
}
\arguments{
\item{mom}{a list of mom's haplotype.}

\item{co_mean}{mean number of crossovers per chromosome.}
}
\value{
a vector of diploid genotype.
}
\description{
Returns a list of true [[1]] and observed [[2]] kid
}
\examples{
copy.mom(list(c(0, 1, 1, 0, 1, 1)), 0.25)
}

