% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lassoC}
\alias{lassoC}
\title{logistic regression with lasso for imputation}
\usage{
lassoC(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link{impute}} function
}
\description{
logistic regression with lasso for imputation
}
\examples{
data(spect)
missdata <- SimIm(spect, 0.1)
\dontrun{
impdata <- impute(spect, cFun = "lassoC")
}
}
\seealso{
\code{\link{cv.glmnet}} and \code{\link{glmnet}}
}

