% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{lassoR}
\alias{lassoR}
\title{LASSO for regression}
\usage{
lassoR(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link[imputeR]{impute}} function
}
\description{
LASSO variable selection for continuous data
}
\examples{
data(parkinson)
missdata <- SimIm(parkinson, 0.1)
\donttest{
impdata <- impute(missdata, lmFun = "lassoR")
}
}
