% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loading_hdf5.R
\name{loading_hdf5}
\alias{loading_hdf5}
\title{\code{Load HDF5 file and recode the genotype matrix. }}
\usage{
loading_hdf5(h5file = "largedata/teo.h5")
}
\arguments{
\item{h5file}{h5 file from tassel. For example, "teo.h5",}

\item{save.file}{RData object to save. For example, "out.RData".}
}
\value{
return three R objects: "genos", "info", "imiss". "genos", genotype matrix. "info", loci information, including missing
rate, maf, snp position, ref and alt alleles, etc. "imiss", individual plant missing rate.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
Load HDF5 file and recode into `0, 1, 2, 3` format. In the genotype file, `0, 1, 2` indicate the copy of alternative allele,
`3` indicate missing data. This function also calculate individual and locus missing rate.
}
\examples{
# note this function requires at least 100G memeory to load teo.h5 file.
obj <- loading_hdf5(h5file="largedata/teo.h5")
}

