\name{mr}
\alias{mr}
\title{calculate miss-classification error}
\usage{
mr(imp, mis, true)
}
\arguments{
  \item{imp}{the imputaed data matrix}

  \item{mis}{the missing data matrix}

  \item{true}{the ture data matrix}
}
\description{
This function calculates the misclassfication error given
the imputed data, the missing data and the true data.
}

