% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imputeKids.R
\name{optimal_win}
\alias{get_break_point}
\alias{hap_in_chunk}
\alias{optimal_win}
\alias{refine_brk_point}
\alias{which_kid_hap}
\title{\code{Maximum likelihood of the inherited haplotypes}}
\usage{
optimal_win(momphase, kid, winstart, winend, stepsize, expect_recomb)

hap_in_chunk(momphase, c, win_length, kid)

which_kid_hap(haplotype, kidwin)

refine_brk_point(mychunk, win_length, verbose, kid)

get_break_point(mychunk)
}
\arguments{
\item{winstart}{The min window size for the heterozygote sites of the two parents.}

\item{winend}{The max window size for the heterozygote sites of the two parents.}

\item{stepsize}{The step size.}

\item{kidwin}{The genotypes of the kid for all heterozygote site of both parents.}

\item{verbose}{Writing verbose messages. Default=TRUE.}

\item{GBS.array}{A GBS.array object with phased parents.
It can be generated from \code{phase_parent} or\code{sim.array} functions.}

\item{join_length}{The length of each neighboring chunks used to connect them into a longer one.}

\item{p1_haps}{Two haplotypes in a chunk of the first parent.}

\item{p2_haps}{All possible haplotypes in chunks of the 2nd parent.}
}
\value{
return a data.frame with all the log likelihoods. Using function \code{momgeno} to extract mom's genotype.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
Determining the optimal window size by searching from winstart to winend.
And then calculate the maximum likelihood of the inherited haplotypes in the determined window size.
}
\examples{
GBS.array <- sim.array(size.array=50, numloci=1000, imiss=0.2, selfing=0.5)
}

