% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pcrR}
\alias{pcrR}
\title{Principle component regression for imputation}
\usage{
pcrR(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link[imputeR]{impute}} function
}
\description{
Principle component regression method for imputation
}
\examples{
data(parkinson)
missdata <- SimIm(parkinson, 0.1)
\donttest{
impdata <- impute(missdata, lmFun = "pcrR")
}
}
\seealso{
\code{\link[pls]{pcr}}
}
