% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PhaseParent.R
\name{phase_parent}
\alias{phase_parent}
\title{\code{Phasing haplotypes of focal parent}}
\usage{
phase_parent(GBS.array, win_length = 10, join_len = 10, hom.error = 0.02,
  het.error = 0.8, imiss = 0.2, verbose = TRUE, unphased_mom = FALSE)
}
\arguments{
\item{GBS.array}{A GBS.array object generated from create_GBS.array() or sim.array() functions.}

\item{win_length}{A list of vectors of Kid's GBS data. Coded with 0, 1 and 2, which is the copy of alternative alleles.
Missing data should be coded with 3.}

\item{join_len}{}

\item{hom.error}{Homozygous error rate, default=0.02. Used for construction of error metrics.}

\item{het.error}{Heterozygous error rate, default=0.8. Used for construction of error metrics.}

\item{imiss}{The cutoff used for determining the log likelihood ratio of the highest and the 2nd highest genotypes.
The oddratio = NULL means to report the most likely genotype. Default value sets to 0.5.}

\item{verbose}{For function 'genomom', returnall=TRUE will return with all the information; returnall=FALSE will return
mom's genotype only.}
}
\value{
return a data.frame with all the log likelihoods. Using function \code{momgeno} to extract mom's genotype.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
Phasing haplotypes using parents' genotype data and progeny array's genotypic data.
Parents' genotypic data can be either confident genotyping data, i.e., WGS data, or
imputed genotypic data generated by \code{impute_parent}. This function was composed of two major parts:
\code{phase_chunk} and \code{join_chunks}.
}
\examples{
obs_mom <- c(0, 0, 0)
obs_kids <- list(c(0, 0, 0), c(0, 0, 0), c(0, 0, 0), c(0, 0, 0), c(0, 0, 0), c(0, 0, 0),
c(0, 0, 0), c(0, 0, 0), c(0, 0, 0), c(0, 0, 0), c(0, 0, 0), c(0, 0, 0))

impute_parent(obs_mom, obs_kids, hom.error=0.02, het.error=0.8)

GBS.array <- sim.array(size.array=50, numloci=1000)
}

