\name{plotIm}
\alias{plotIm}
\title{Plot function for imputation}
\usage{
plotIm(imp, mis, true, ...)
}
\arguments{
  \item{imp}{the imputed data matrix}

  \item{mis}{the missing data matrix}

  \item{true,}{the true data matrix}
}
\description{
this is a plot function for assessing imputation
performance given the imputed data and the original true
data
}

