% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{ran.hap}
\alias{ran.hap}
\title{Create random haplotype with sfs}
\usage{
ran.hap(numloci, p)
}
\arguments{
\item{numloci}{number of loci.}

\item{p}{a vector of allele freq with length numloci.}
}
\value{
a vector of genotype,
}
\description{
Create random haplotype with sfs
}
\examples{
ran.hap(3, c(0.5, 0.8, 0.1))
}

