% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rdaC}
\alias{rdaC}
\title{regularised LDA method for imputation}
\usage{
rdaC(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link{impute}} function
}
\description{
regularised LDA method for imputation
}
\examples{
data(spect)
missdata <- SimIm(spect, 0.1)
\dontrun{
impdata <- impute(spect, cFun = "rdaC")
}
}
\seealso{
\code{\link{rda}}
}
