% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ridgeC}
\alias{ridgeC}
\title{Ridge regression with lasso for imputation}
\usage{
ridgeC(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link[imputeR]{impute}} function
}
\description{
Ridge regression with lasso for imputation
}
\examples{
data(spect)
missdata <- SimIm(spect, 0.1)
\donttest{
impdata <- impute(spect, cFun = "ridgeC")
}
}
\seealso{
\code{\link[ridge]{logisticRidge}}
}
