% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ridgeR}
\alias{ridgeR}
\title{Ridge shrinkage for regression}
\usage{
ridgeR(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link{impute}} function
}
\description{
Ridge shrinkage variable selection for continuous data
}
\examples{
data(parkinson)
missdata <- SimIm(parkinson, 0.1)
\dontrun{
impdata <- impute(missdata, lmFun = "ridgeR")
}
}
