% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rpartC}
\alias{rpartC}
\title{classification tree for imputation}
\usage{
rpartC(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link{impute}} function
}
\description{
classification tree for imputation
}
\examples{
data(spect)
missdata <- SimIm(spect, 0.1)
\dontrun{
impdata <- impute(spect, cFun = "rpartC")
}
}
\seealso{
\code{\link{rpart}}
}
