% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.array.R
\name{sim.array}
\alias{kid}
\alias{sim.array}
\title{\code{A progeny array simulator.}}
\usage{
sim.array(size.array, numloci, hom.error = 0.02, het.error = 0.8,
  rec = 0.25, selfing = 0.5, imiss = 0.5, misscode = 3)

kid(p1, p2, het.error, hom.error, rec, imiss, misscode)
}
\arguments{
\item{size.array}{Size of the progeny array.}

\item{numloci}{Number of loci to simulate.}

\item{hom.error}{Homozygous error rate, default=0.02.}

\item{het.error}{Heterozygous error rate, default=0.8.}

\item{rec}{Recombination rate, default=0.25.}

\item{selfing}{Proportion of selfed progeny, default=0.5.}

\item{imiss}{Individual missing rate, default=0.5.}

\item{misscode}{Missing code, default=3.}

\item{p1}{first parent (a list)}

\item{p2}{second parent (a list)}
}
\value{
Return GBS.array object.
Slot1: true_parents, a list of data.frame(hap1, hap2).
Slot2: gbs_parents, a list of genotypes. For example, c(1, 2, 2, 0, 3).
Slot3: true_kids, a list of data.frame(hap1, hap2).
Slot4: gbs_kids, a list of kid genotypes. For example, c(1, 1, 3, 1, 2).
Slot5: pedigree, a data.frame (kid, p1, p2). Note, p1 is the focal parent.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.

a list of true [[1]] and observed [[2]] kid
}
\description{
Simulate a S4 GBS.array object with true_parents, gbs_parents, true_kids, gbs_kids and pedigree.
}
\examples{
test <- sim.array(size.array=50, numloci=10)
class(test)
}

