% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimSelfer_OXer.R
\name{SimSelfer}
\alias{SimOXer}
\alias{SimSelfer}
\alias{add_error}
\alias{copy.mom}
\alias{kid}
\alias{missing.idx}
\title{\code{Simulate half sib and full sib GBS data. }}
\usage{
SimSelfer(size.array = 20, het.error = 0.8, hom.error = 0.002,
  numloci = 1000, rec = 1.5, imiss = 0.3)

SimOXer(size.array = 10, het.error = 0.7, hom.error = 0.002,
  numloci = 1000, rec = 1.5, imiss = 0.3, misscode = 3)

add_error(diploid, hom.error, het.error)

copy.mom(mom, co_mean)

missing.idx(nloci, imiss)

kid(mom, dad, het.error, hom.error, rec = 1.5, imiss = 0.3, misscode = 3)
}
\arguments{
\item{size.array}{Size of the progeny array.}

\item{het.error}{Heterozygous error rate, default=0.8.}

\item{hom.error}{Homozygous error rate, default=0.002.}

\item{numloci}{Number of loci to simulate.}

\item{rec}{Recombination rate, default=1.5.}

\item{imiss}{Individual missing rate.}
}
\value{
return a list of two. The first is a vector of mom's genotype; the second is a list of vectors
of kids' genotype.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
Simulate half sib's GBS genotype using 'SimSelfer',
Simulate full sib's GBS genotype using 'SimOXer'.
size.array=20, het.error=0.7, hom.error=0.002, numloci=1000, rec=1.5, imiss=0.3
}
\examples{
SimSelfer(size.array=20, het.error=0.8, hom.error=0.002, numloci=1000, rec=1.5, imiss=0.3)

momgeno(geno, oddratio=0.5, returnall=FALSE)
}

