% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stepBackR}
\alias{stepBackR}
\title{Best subset (backward direction) for regression}
\usage{
stepBackR(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\description{
Best subset variable selection (backward direction) for continuous data
}

