% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stepBothC}
\alias{stepBothC}
\title{Best subset for classification (both direction)}
\usage{
stepBothC(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\description{
Best subset variable selection from both forward and backward
direction for categorical data
}

