% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stepForR}
\alias{stepForR}
\title{Best subset (forward direction) for regression}
\usage{
stepForR(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\description{
Best subset variable selection (forward direction) for continuous data
}

