% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TasselHDF5-class}
\alias{TasselHDF5-class}
\title{An S4 class that stores links to Tassel HDF5 objects, loci positions, and genotypes}
\description{
An S4 class that stores links to Tassel HDF5 objects, loci positions, and genotypes
}
\section{Slots}{

\describe{
\item{\code{filename}}{path to HDF5 file}

\item{\code{ranges}}{a \code{GenomicRanges} object of loci}

\item{\code{ref}}{a Tassel-encoded integer vector of reference alleles}

\item{\code{alt}}{a Tassel-encoded \code{IntegerList} of alternate alleles}

\item{\code{genotypes}}{a matrix of bialleic genotypes}

\item{\code{samples}}{sample names}

\item{\code{version}}{string Tassel version}
}}

