% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_array.R
\name{update_gbs_parents}
\alias{update_gbs_parents}
\title{\code{Update GBS.array object}}
\usage{
update_gbs_parents(GBS.array, ipdat)
}
\arguments{
\item{GBS.array}{input object.}

\item{ipdat}{a data.frame of imputed gbs parents.}
}
\value{
Return updated GBS.array object. 

Details of the GBS.array objects.
Slot1: true_parents, a list of data.frame(hap1, hap2).
Slot2: gbs_parents, a list of genotypes. For example, c(1, 2, 2, 0, 3).
Slot3: true_kids, a list of data.frame(hap1, hap2).
Slot4: gbs_kids, a list of kid genotypes. For example, c(1, 1, 3, 1, 2).
Slot5: pedigree, a data.frame (kid, p1, p2). Note, p1 is the focal parent.
Slot6: freq, a vector of reference allele freq for all SNPs.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
Update GBS.array object slot gbs_parents.
}
\examples{
update_gbs_parents(GBS.array, ipdat)

}

