% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_array.R
\name{create_array}
\alias{create_array}
\alias{est_frq}
\alias{get_array_item}
\alias{ped_focal}
\alias{pedinfo}
\title{\code{Create GBS.array objects}}
\usage{
create_array(Geno4imputeR, ped, outdir = "largedata/", maf_cutoff = 0.002,
  lmiss_cutoff = 0.8, imiss_cutoff = 0.8, size_cutoff = 40)

pedinfo(ped)

est_frq(pgeno)

ped_focal(ped, focalp)

get_array_item(info, geno, myped, chrj, focalp)
}
\arguments{
\item{Geno4imputeR}{input object.}

\item{ped}{a data.frame of pedigree information, must include three columns: proid (progeny id),
parent1 (the first parent id), parent2 (the 2nd parent id).}

\item{outdir}{path of the output dir.}

\item{maf_cutoff}{cutoff for MAF, default=0.002.}

\item{lmiss_cutoff}{cutoff for locus missing rate, default=0.8.}

\item{imiss_cutoff}{cutoff for individual missing rate, default=0.8.}

\item{size_cutoff}{Minimum family size required for imputation, default=40.}
}
\value{
Return updated Geno4imputeR object.
The GBS.array objects will be output to the given direction by family and chr.

Details of the GBS.array objects.
Slot1: true_parents, a list of data.frame(hap1, hap2).
Slot2: gbs_parents, a list of genotypes. For example, c(1, 2, 2, 0, 3).
Slot3: true_kids, a list of data.frame(hap1, hap2).
Slot4: gbs_kids, a list of kid genotypes. For example, c(1, 1, 3, 1, 2).
Slot5: pedigree, a data.frame (kid, p1, p2). Note, p1 is the focal parent.
Slot6: freq, a vector of reference allele freq for all SNPs.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
Create GBS.array objects from Geno4imputeR objects. It will compute population allele frq based on families
of selfed kids. Original Geno4imputeR will be updated and returned. A set of GBS.array objects will be output
to the directory by chromosomes.
}
\examples{
create_array(Geno4imputeR, ped, outdir="largedata/",
maf_cutoff=0.002, lmiss_cutoff=0.8, imiss_cutoff=0.8, size_cutoff=40)
}

