% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GBS.array-class}
\alias{GBS.array-class}
\title{An S4 class that stores links to Geno4imputeR objects, genotypes, loci missing rate, and MAF}
\description{
An S4 class that stores links to Geno4imputeR objects, genotypes, loci missing rate, and MAF
}
\section{Slots}{

\describe{
\item{\code{genomx}}{a matrix of bialleic genotypes, 0 ref, 1 het, 2 alt and 3 missing}

\item{\code{info}}{SNP information, including SNP ID, chr, position, ref and alt alleles, loci missing rate.}

\item{\code{imiss}}{individual plant missing rate}
}}

