% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImputeMom.R
\name{impute_mom}
\alias{impute_mom}
\title{\code{Impute mom's genotype using mom and kids' GBS data. }}
\usage{
impute_mom(obs_mom, obs_kids, locus, p)
}
\arguments{
\item{obs_mom}{A vector of mom's GBS data. Coded with 0, 1 and 2, which is the copy of alternative alleles.
Missing data should be coded with 3.}

\item{obs_kids}{A list of vectors of Kid's GBS data. Coded with 0, 1 and 2, which is the copy of alternative alleles.
Missing data should be coded with 3.}

\item{p}{p=sample(sfs,numloci) #freqs of all loci sampled from site frequency spectrum (SFS)}
}
\value{
If all inputs are integer and logical, then the output
  will be an integer. If integer overflow
  \url{http://en.wikipedia.org/wiki/Integer_overflow} occurs, the output
  will be NA with a warning. Otherwise it will be a length-one numeric or
  complex vector.

  Zero-length vectors have sum 0 by definition. See
  \url{http://en.wikipedia.org/wiki/Empty_sum} for more details.
}
\description{
We have obs. mom and obs. (selfed) kids. We want to know the likelihood of the mom's genotypes.
This function is to impute mom's most likely genotype from a progeny array of k kids by giving a log likelihood threshold.
inferred_mom=1 -> 00, 2->01, 3->11
}
\examples{
obs_mom <- c(0, 0, 0)
obs_kids <- list(c(1, 1, 1), c(0, 0, 0), c(1, 1, 1))

sum(.Machine$integer.max, 1L)
sum(.Machine$integer.max, 1)
}

