% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImputeParent.R
\name{impute_parent}
\alias{impute_one_site}
\alias{impute_parent}
\alias{parentgeno}
\title{\code{Impute parent's genotype using parent and kids' GBS data. }}
\usage{
impute_parent(parents, obs_parent, other_parents, obs_kids, hom.error = 0.02,
  het.error = 0.8, p = NULL)

impute_one_site(locus, gen_error, p_locus, probs, parents, obs_parent,
  other_parents, obs_kids)

parentgeno(geno, oddratio = 0.6931472, returnall = TRUE)
}
\arguments{
\item{parents}{A list of the genotypes of all possible parents, including current parent. These are observed GBS data, coded with 0, 1 and 2, which is the copy of alternative alleles.
Missing data should be coded with 3.}

\item{obs_parent}{An integer referring to which parent is the current parent of interest.}

\item{other_parents}{A vector of integers referring to which are the second parent of each kid.
Should be same length as number of kids (length(obs_kids))}

\item{obs_kids}{A list of vectors of Kid's GBS data. Coded with 0, 1 and 2, which is the copy of alternative alleles.
Missing data should be coded with 3.}

\item{hom.error}{Homozygous error rate, default=0.02.}

\item{het.error}{Heterozygous error rate, default=0.8.}

\item{p}{vector of allele frequencies at each locus. should be estimated from set of all parents.
WARNING: if not supplied, a random neutral SFS is used to generate p at each locus.}

\item{oddratio}{The cutoff used for determining the log likelihood ratio of the highest and the 2nd highest genotypes.
The oddratio = NULL means to report the most likely genotype. Default value sets to 0.6931472,
which is equivalent to most likely genotype being twice as likely as next most likely. 2X as likely}

\item{returnall}{For function 'parentgeno', returnall=TRUE will return with all the information; returnall=FALSE will return
parent's genotype only (either maxlikelihood or, if oddratio is specified, genotypes with low oddratios will be set to missing).}
}
\value{
return a data.frame with all the log likelihoods. Using function \code{parentgeno} to extract parent's genotype.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
We have observed parent and observed kids. We want to know the likelihood of the parent's possible genotypes.
This function is to impute the parent's most likely genotype from a progeny array of k kids by
giving a log likelihood threshold.
}
\examples{
obs_parent <- c(0, 0, 0)
obs_kids <- list(c(0, 0, 0), c(0, 0, 0), c(0, 0, 0), c(0, 0, 0), c(0, 0, 0), c(0, 0, 0),
c(0, 0, 0), c(0, 0, 0), c(0, 0, 0), c(0, 0, 0), c(0, 0, 0), c(0, 0, 0))

impute_parent(obs_parent, obs_kids, hom.error=0.02, het.error=0.8)

parentgeno(geno, oddratio=0.5, returnall=FALSE)
}

