% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeR-package.R
\docType{package}
\name{imputeR-package}
\alias{imputeR-package}
\title{imputeR-package description}
\description{
The imputeR package offers a General Multivariate Imputation Framework
}
\details{
The imputeR package is a Multivariate Expectation-Maximization (EM) based imputation framework that offers several
different algorithms. These include regularisation methods like Lasso and Ridge regression, tree-based models and dimensionality
reduction methods like PCA and PLS.
}
\author{
Steffen Moritz, Lingbing Feng, Gen Nowak, Alan. H. Welsh, Terry. J. O'Neill
}
