% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeR-package.R
\docType{package}
\name{imputeR-package}
\alias{imputeR-package}
\title{imputeR-package description}
\description{
The imputeR Package offers a General Imputation Framework for R
}
\details{
This package provides a general imputation framework based on
variable selection methods including regularisation methods,
tree-based models and dimension reduction methods.
}
\author{
Steffen Moritz, Lingbing Feng, Gen Nowak, Alan. H. Welsh, Terry. J. O'Neill
}
