% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my-TasselHDF5-method.R
\name{imputeRob}
\alias{imputeRob}
\title{\code{Create TasselHDF5 object. }}
\usage{
imputeRob(teo)
}
\arguments{
\item{inpob}{A S4 method for TasselHDF5 object.}
}
\value{
return "Geno4imputeR" object with three slots: "genomx", "info", "imiss".
"genomx", genotype matrix.
"info", loci information, including SNP ID (snpid), chr, position (pos),
loci missing rate (lmiss), minor allele freq (maf).
"imiss", individual plant missing rate.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
S4 method for TasselHDF5 object. Recode it into `0, 1, 2, 3` format.
In the genotype file, `0, 1, 2` indicate the copy of alternative allele,
`3` indicate missing data. This function also calculate individual and locus missing rate.
}
\examples{
# note this function requires at least 100G memeory to load teo.h5 file.
obj <- loading_hdf5(h5file="largedata/teo.h5")
}

