% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{major}
\alias{major}
\title{Majority imputation for a vector}
\usage{
major(x)
}
\arguments{
\item{x}{a character (or numeric categorical) vector with missing values}
}
\value{
the same length of vector with missing values being imputed by the majority class
in this vector.
}
\description{
This function is internally used by \code{\link{guess}}, it 
may be useless in reality.
}
\examples{
a <- c(rep(0, 10), rep(1, 15), rep(2, 5))
a[sample(seq_along(a), 5)] <- NA
a
b <- major(a)
b
}
