% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mr}
\alias{mr}
\title{calculate miss-classification error}
\usage{
mr(imp, mis, true)
}
\arguments{
\item{imp}{the imputaed data matrix}

\item{mis}{the missing data matrix}

\item{true}{the ture data matrix}
}
\value{
The missclassification error
}
\description{
This function calculates the misclassfication error given the imputed data,
the missing data and the true data.
}
\examples{
data(spect)
Detect(spect)
missdata <- SimIm(spect, 0.1)
\dontrun{
require(cutoffR)
HeatStruct(missdata)
nmissing(missdata)
# impute using rpart
impdata <- impute(missdata, cFun = "rpartC")
# calculate the misclassification error
mr(impdata$imp, missdata, spect)
}
}

