% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PhaseParent.R
\name{phase_chunk}
\alias{infer_dip}
\alias{jump_win}
\alias{maxlog_hap_selfed_kid}
\alias{phase_chunk}
\title{\code{Phasing Parent's haplotypes into chunks. }}
\usage{
phase_chunk(GBS.array, haps, verbose, win_length)

infer_dip(GBS.array, winidx, haps, returnhap = FALSE)

maxlog_hap_selfed_kid(haplotype, kid_geno)

jump_win(GBS.array, winstart, win_length, hetsites, haps)
}
\arguments{
\item{GBS.array}{A GBS.array object generated from create_GBS.array() or sim.array() functions.}

\item{haps}{all possible haplotypes}

\item{verbose}{For function 'genomom', returnall=TRUE will return with all the information; returnall=FALSE will return
mom's genotype only.}

\item{win_length}{A list of vectors of Kid's GBS data. Coded with 0, 1 and 2, which is the copy of alternative alleles.
Missing data should be coded with 3.}
}
\value{
return a data.frame with all the log likelihoods. Using function \code{momgeno} to extract mom's genotype.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
Phasing haplotypes using parents' genotype data and progeny array's genotypic data.
Parents' genotypic data can be either confident genotyping data, i.e., WGS data, or
imputed genotypic data generated by \code{impute_parent}.
}
\examples{
GBS.array <- sim.array(size.array=50, numloci=1000)
haps <- setup_haps(win_length)
haplist <- phase_chuck(GBS.array, haps, verbose, win_length)
}

