% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PhaseParent.R
\name{phase_parent}
\alias{phase_error_rate}
\alias{phase_parent}
\alias{write_phase}
\title{\code{Phasing haplotypes of focal parent}}
\usage{
phase_parent(GBS.array, win_length = 10, join_length = 10,
  self_cutoff = 20, verbose = TRUE)

write_phase(outchunks)

phase_error_rate(GBS.array, phase)
}
\arguments{
\item{GBS.array}{A GBS.array object generated from \code{impute_parent} or \code{sim.array} functions.
The slot \code{gbs_parents} replaced with imputed parental genotypes.}

\item{win_length}{Window length used for halotype phasing. Default=10.
win_length > 20 will dramatically increase computational burden.}

\item{join_length}{The length of each neighboring chunks used to connect them into a longer one.}

\item{verbose}{Writing verbose messages. Default=TRUE.}
}
\value{
return a data.frame with all the log likelihoods. Using function \code{momgeno} to extract mom's genotype.

  See \url{https://github.com/yangjl/imputeR/blob/master/vignettes/imputeR-vignette.pdf} for more details.
}
\description{
Phasing focal parent's haplotypes using parents' genotypic data and progeny array's genotypic data.
Parents' genotypic data can be either confident genotyping data, i.e., WGS data, or
imputed genotypic data generated by \code{impute_parent}. This function was composed of two major parts:
\code{phase_chunk} and \code{join_chunks}. The haplotype was phased using a sliding window and extended to
a large chunk of sequence. The chunks were then joined into chromosomal wide haplotypes.
}
\examples{
GBS.array <- sim.array(size.array=50, numloci=1000, imiss=0.2, selfing=0.5)
}

