% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plsR}
\alias{plsR}
\title{Partial Least Square regression for imputation}
\usage{
plsR(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link{impute}} function
}
\description{
Principle component regression method for imputation
}
\examples{
data(parkinson)
missdata <- SimIm(parkinson, 0.1)
\dontrun{
impdata <- impute(missdata, lmFun = "plsR")
}
}
\seealso{
\code{\link{plsr}}
}

