% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ridgeC}
\alias{ridgeC}
\title{Ridge regression with lasso for imputation}
\usage{
ridgeC(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link{impute}} function
}
\description{
Ridge regression with lasso for imputation
}
\examples{
data(spect)
missdata <- SimIm(spect, 0.1)
\dontrun{
impdata <- impute(spect, cFun = "ridgeC")
}
}
\seealso{
\code{\link{logisticRidge}}
}

