% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stepBackC}
\alias{stepBackC}
\title{Best subset for classification (backward)}
\usage{
stepBackC(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link{impute}} function
}
\description{
Best subset variable selection from both forward and backward
direction for categorical data
}
\examples{
data(spect)
missdata <- SimIm(spect, 0.1)
\dontrun{
impdata <- impute(spect, cFun = "stepBackC")
}
}
\seealso{
\code{\link{step}}, \code{\link{stepBackR}}
}
