% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stepBothC}
\alias{stepBothC}
\title{Best subset for classification (both direction)}
\usage{
stepBothC(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link{impute}} function
}
\description{
Best subset variable selection from both forward and backward
direction for categorical data
}
\examples{
data(spect)
missdata <- SimIm(spect, 0.1)
\dontrun{
impdata <- impute(spect, cFun = "stepBothC")
}
}
\seealso{
\code{\link{step}}, \code{\link{stepBothR}}
}

