% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stepBothR}
\alias{stepBothR}
\title{Best subset for regression (both direction)}
\usage{
stepBothR(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link{impute}} function
}
\description{
Best subset variable selection from both forward and backward
direction for continuous data
}
\examples{
data(parkinson)
missdata <- SimIm(parkinson, 0.1)
\dontrun{
impdata <- impute(missdata, lmFun = "stepBothR")
}
}

