% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stepBothR}
\alias{stepBothR}
\title{Best subset for regression (both direction)}
\usage{
stepBothR(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link[imputeR]{impute}} function
}
\description{
Best subset variable selection from both forward and backward
direction for continuous data
}
\examples{
data(parkinson)
missdata <- SimIm(parkinson, 0.1)
\donttest{
impdata <- impute(missdata, lmFun = "stepBothR")
}
}
