% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stepForC}
\alias{stepForC}
\title{Best subset for classification (forward direction)}
\usage{
stepForC(x, y)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}
}
\value{
a model object that can be used by the \code{\link[imputeR]{impute}} function
}
\description{
Best subset variable selection from both forward and backward
direction for categorical data
}
\examples{
data(spect)
missdata <- SimIm(spect, 0.1)
\donttest{
impdata <- impute(spect, cFun = "stepForC")
}
}
\seealso{
\code{\link[stats]{step}}, \code{\link[imputeR]{stepForR}}
}
